Moved parse-related stuff into a new proxyp/Parser.h
